+++ global.d.ts
@@ -1,0 +1,11 @@
+namespace global {
+  declare module "*.module.scss" {
+    const content: { [className: string]: string };
+    export default content;
+  }
+
+  declare module "*.module.sass" {
+    const content: { [className: string]: string };
+    export default content;
+  }
+}