+++ packages/form-components/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./components";
 export * from "./auth";
+export * from "./actions";
+export * from "./item-select";